*** empty log message ***
git-svn-id: http://svn.vuze.com/public/client@6006 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/core/dht/DHTStorageAdapter.java
@@ -83,5 +83,5 @@ DHTStorageAdapter
 	createNewDiversification(
 		byte[]			key,
 		boolean			put_operation,
-		int				diversification_type );
+		byte			diversification_type );
 }
